Simplify implementation and avoid code duplication
