+++ src/components/smart-grid/internal/smart-grid-virtual-scroller/get-new-start-and-end-indexes.ts
@@ -1,0 +1,212 @@
+import { inBetween } from "../../../../common/utils";
+import { SmartGridModel } from "../../types";
+import { SmartGridCellsToLoad, SmartGridCellVirtualSize } from "./types";
+import {
+  cellIsRendered,
+  getSmartCells,
+  isRenderedSmartCellVisible,
+  isVirtualSizeCellVisible
+} from "./utils";
+
+const CAN_NOT_CHECK_SHIFT_VALUES = {
+  type: "waiting-for-cells-to-render"
+} as const satisfies SmartGridCellsToLoad;
+
+const findFirstVirtualSizeThatIsNotVisible = (
+  renderedCells: HTMLChSmartGridCellElement[],
+  items: SmartGridModel,
+  virtualSizes: Map<string, SmartGridCellVirtualSize>,
+  bufferSize: number,
+  smartGridScrollTop: number,
+  smartGridBoundingRect: DOMRect
+): SmartGridCellsToLoad => {
+  let closerVirtualSizeId = "";
+  let closerVirtualSizeBottomY = 0;
+
+  // Find the key of the closer virtual size to the viewport that isn't visible
+  virtualSizes.forEach((virtualSize, virtualSizeId) => {
+    const virtualSizeBottomY = virtualSize.offsetTop + virtualSize.height;
+    const virtualSizeIsHidden = virtualSizeBottomY <= smartGridScrollTop;
+
+    if (virtualSizeIsHidden && closerVirtualSizeBottomY <= virtualSizeBottomY) {
+      closerVirtualSizeBottomY = virtualSizeBottomY;
+      closerVirtualSizeId = virtualSizeId;
+    }
+  });
+
+  // TODO: Use memory to retrieve the index given the cellId
+  const closerVirtualItemIndex = items.findIndex(
+    el => el.id === closerVirtualSizeId
+  );
+
+  const startIndex = Math.max(0, closerVirtualItemIndex - bufferSize - 1);
+  const lastIndex = items.length - 1;
+  let endIndex = closerVirtualItemIndex + 1; // Start in the first visible cell
+  let renderedCellsCount = 0;
+  let cellsThatAreNotVisible = 0;
+
+  // Find the endIndex to render the cells. This index takes into account the
+  // cells that must be not visible in the buffer
+  while (
+    inBetween(0, endIndex, lastIndex) &&
+    cellsThatAreNotVisible < bufferSize
+  ) {
+    const cellId = items[endIndex].id;
+
+    const virtualSize = virtualSizes.get(cellId);
+
+    if (virtualSize) {
+      const isVirtualSizeVisible = isVirtualSizeCellVisible(
+        virtualSize,
+        smartGridScrollTop,
+        smartGridBoundingRect
+      );
+
+      if (!isVirtualSizeVisible) {
+        cellsThatAreNotVisible++;
+      }
+    }
+    // We assume that the rendered cells are sorted
+    else {
+      const renderedCell = renderedCells[renderedCellsCount];
+
+      const isRenderedCellVisible = isRenderedSmartCellVisible(
+        renderedCell,
+        smartGridBoundingRect
+      );
+
+      if (!isRenderedCellVisible) {
+        cellsThatAreNotVisible++;
+      }
+
+      renderedCellsCount++;
+    }
+
+    endIndex++;
+  }
+
+  endIndex = Math.min(lastIndex, endIndex + bufferSize - 1);
+
+  return {
+    startIndex,
+    endIndex,
+    renderedCells,
+    newRenderedCellStartIndex: renderedCellsCount === 0 ? -1 : 0,
+    newRenderedCellEndIndex: renderedCellsCount,
+    type: "index"
+  };
+};
+
+const LOWER_BUFFER_RANGE = (index: number, firstIndex) => firstIndex <= index;
+const UPPER_BUFFER_RANGE = (index: number, lastIndex) => index <= lastIndex;
+
+const getAmountOfCellsThatAreVisibleInTheBuffer = (
+  increment: -1 | 1,
+  bufferSize: number,
+  renderedCells: HTMLChSmartGridCellElement[],
+  smartGridBoundingRect: DOMRect
+): number => {
+  const lastRenderedCellIndex = renderedCells.length - 1;
+
+  const indexLimit =
+    increment === 1 ? lastRenderedCellIndex - bufferSize : bufferSize;
+  const checkIndexRange =
+    increment === 1 ? UPPER_BUFFER_RANGE : LOWER_BUFFER_RANGE;
+
+  // Depending on the direction, start from the first or last index of the
+  // current rendered cells
+  let currentIndex = increment === 1 ? 0 : lastRenderedCellIndex;
+  let shift = bufferSize;
+
+  // Check how many cells are visible in the buffer
+  while (checkIndexRange(currentIndex, indexLimit)) {
+    const currentCell = renderedCells[currentIndex];
+
+    if (isRenderedSmartCellVisible(currentCell, smartGridBoundingRect)) {
+      break;
+    }
+
+    shift--;
+    currentIndex += increment;
+  }
+
+  return shift;
+};
+
+/**
+ * Depending on the scroll position and the bufferSize, it returns the start
+ * and end index of the virtual array to rendered the cells
+ */
+export const getNewStartAndEndIndexes = (
+  scroller: HTMLChSmartGridVirtualScrollerElement,
+  smartGrid: HTMLChSmartGridElement,
+  items: SmartGridModel,
+  virtualSizes: Map<string, SmartGridCellVirtualSize>,
+  virtualStartSize: number,
+  virtualEndSize: number,
+  bufferSize: number
+): SmartGridCellsToLoad => {
+  const renderedCells = getSmartCells(scroller);
+  const allCellsAreRendered = renderedCells.every(cellIsRendered);
+
+  // All cells must be rendered before trying to update the DOM
+  if (!allCellsAreRendered) {
+    return CAN_NOT_CHECK_SHIFT_VALUES;
+  }
+
+  // DOM read operations
+  const smartGridBoundingRect = smartGrid.getBoundingClientRect();
+  const smartGridScrollTop = smartGrid.scrollTop;
+
+  // TODO: Force the buffer to be at least 2
+  const secondRenderedCell = renderedCells[1];
+
+  const scrollIsAtVirtualStartSize =
+    virtualStartSize > 0 && smartGridScrollTop < secondRenderedCell.offsetTop;
+
+  if (scrollIsAtVirtualStartSize) {
+    return findFirstVirtualSizeThatIsNotVisible(
+      renderedCells,
+      items,
+      virtualSizes,
+      bufferSize,
+      smartGridScrollTop,
+      smartGridBoundingRect
+    );
+  }
+
+  const secondLastCell = renderedCells.at(-2);
+
+  const scrollIsAtVirtualEndSize =
+    virtualEndSize > 0 &&
+    secondLastCell.offsetTop + secondLastCell.getBoundingClientRect().height <
+      smartGridScrollTop;
+
+  if (scrollIsAtVirtualEndSize) {
+    console.log(
+      "///////////VIRTUAL END SIZE///////////",
+      "cellId " + secondLastCell.cellId,
+
+      "secondLastCell.bottomY " +
+        (secondLastCell.offsetTop +
+          secondLastCell.getBoundingClientRect().height),
+
+      "smartGridScrollTop" + smartGridScrollTop
+    );
+  }
+
+  const startShift = getAmountOfCellsThatAreVisibleInTheBuffer(
+    1,
+    bufferSize,
+    renderedCells,
+    smartGridBoundingRect
+  );
+  const endShift = getAmountOfCellsThatAreVisibleInTheBuffer(
+    -1,
+    bufferSize,
+    renderedCells,
+    smartGridBoundingRect
+  );
+
+  return { startShift, endShift, renderedCells, type: "shift" };
+};
